18.9.23 - GameState
18.9.23 - GameState

+++ Assets/Scripts/EnemyHp.cs
@@ -5,9 +5,9 @@ using UnityEngine.UI;
 
 public class EnemyHp : MonoBehaviour
 {
-    public static float maxHp;
-    public static float staticHp;
-    public float hp;
+    public static int maxHp;
+    public static int staticHp;
+    public int hp;
     public Text hpText;
     public Text deckSize;
     public Text deadNo;
